requirements.txt
flask~=2.2.2
GitPython~=3.1.31
WebOb~=1.8.7
flask-http-middleware~=0.2.1
